% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddst.againststochdom.test.R
\name{ddst.againststochdom.test}
\alias{ddst.againststochdom.test}
\title{Data Driven Smooth Test Against Stochastic Dominance}
\usage{
ddst.againststochdom.test(
  x,
  y,
  k.n = 4,
  alpha = 0.05,
  t,
  B = 10000,
  compute.cv = FALSE
)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data}

\item{y}{a (non-empty) numeric vector of data}

\item{k.n}{an integer specifying a level of complexity of the grid considered, only for advanced users}

\item{alpha}{a significance level}

\item{t}{an alpha-dependent tunning parameter in the penalty in the model selection rule}

\item{B}{an integer specifying the number of runs for a p-value and a critical value computation if any}

\item{compute.cv}{a logical value indicating whether to compute a critical value corresponding to the significance level alpha or not}
}
\description{
Performs data driven smooth non-parametric two-sample test against one-sided alternatives
(stochastic dominance).
Suppose that we have random samples from two distributions F and G.
The null hypothesis is that F(x) < G(x) for some x while the alternative is that at
F(x) >= G(x) for all x with strict inequality for at least one x.
Detailed description of the test statistic is provided in Ledwina and Wylupek (2012).
}
\examples{
set.seed(7)
# H0 is true
x <- runif(80)
y <- runif(80)
t <- ddst.againststochdom.test(x, y, alpha = 0.05, t = 2.2, k.n = 4)
t
plot(t)

# H0 is false
# known fixed alternative
x <- runif(80)
y <- rbeta(80,4,2)
t <- ddst.againststochdom.test(x, y, alpha = 0.05, t = 2.2, k.n = 4)
t
plot(t)
}
\references{
Two-sample test against one-sided alternatives. Ledwina and Wylupek (2012). \url{https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1467-9469.2011.00787.x}
}
\keyword{htest}
