% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddst.extr.test.R
\name{ddst.extr.test}
\alias{ddst.extr.test}
\alias{ddst.extr.Nk}
\title{Data Driven Smooth Test for Extreme Value Distribution}
\usage{
ddst.extr.test(x, base = ddst.base.legendre, c = 100, B = 1000,
  compute.p = FALSE, Dmax = 5, ...)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values}

\item{base}{a function which returns orthogonal system, might be \code{ddst.base.legendre} for Legendre polynomials or
\code{ddst.base.cos} for cosine system, see package description}

\item{c}{a parameter for model selection rule, see package description}

\item{B}{an integer specifying the number of replicates used in p-value computation}

\item{compute.p}{a logical value indicating whether to compute a p-value}

\item{Dmax}{an integer specifying the maximum number of coordinates, only for advanced users}

\item{...}{further arguments}
}
\value{
An object of class \code{htest}
  \item{statistic }{the value of the test statistic.}
  \item{parameter }{the number of choosen coordinates (k).}
  \item{method }{a character string indicating the parameters of performed test. }
  \item{data.name }{a character string giving the name(s) of the data. }
  \item{p.value }{the p-value for the test, computed only if \code{compute.p=TRUE}.}
}
\description{
Performs data driven smooth test for composite hypothesis of extreme value distribution.
Null density is given by
\eqn{
  f(z;\gamma)=1/\gamma_2 \exp((z-\gamma_1)/\gamma_2- \exp((z-\gamma_1)/\gamma_2))}, \eqn{z \in R}.
}
\details{
We model alternatives similarly as in Kallenberg and Ledwina (1997) and Janic-Wroblewska (2004) using Legendre's polynomials or cosines.
  For more details see: \url{http://www.biecek.pl/R/ddst/description.pdf}.
}
\examples{
library(evd)

# for given vector of 19 numbers
z <- c(13.41, 6.04, 1.26, 3.67, -4.54, 2.92, 0.44, 12.93, 6.77, 10.09,
      4.10, 4.04, -1.97, 2.17, -5.38, -7.30, 4.75, 5.63, 8.84)
t <- ddst.extr.test(z, compute.p=TRUE)
t
plot(t)

# H0 is true
x <- -qgumbel(runif(100),-1,1)
t <- ddst.extr.test (x, compute.p = TRUE)
t
plot(t)

# H0 is false
x <- rexp(80,4)
t <- ddst.extr.test (x, compute.p = TRUE)
t
plot(t)
}
\references{
Hosking, J.R.M., Wallis, J.R., Wood, E.F. (1985). Estimation of the generalized extreme-value distribution by the method of probability-weighted moments. \eqn{ Technometrics} 27, 251--261.

Janic-Wroblewska, A. (2004). Data-driven smooth test for extreme  value distribution. \eqn{ Statistics} 38, 413--426.

Janic, A. and Ledwina, T. (2008). Data-driven tests for a location-scale family revisited. \eqn{ J. Statist. Theory. Pract. Special issue on Modern Goodness of Fit Methods. accepted.}.

Kallenberg, W.C.M., Ledwina, T. (1997). Data driven smooth tests for composite hypotheses: Comparison of powers. \eqn{ J. Statist. Comput. Simul.} \bold{ 59}, 101--121.
}
\keyword{htest}
