% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddst.ksample.test.R
\name{ddst.ksample.test}
\alias{ddst.ksample.test}
\title{Data Driven k-Sample Test}
\usage{
ddst.ksample.test(x, ..., d_N = 12, c = 2.3, B = 100,
  compute.p = FALSE)
}
\arguments{
\item{x}{a list with vectors or vectors (in thos case other vectors will be in ...)}

\item{...}{other vectors}

\item{d_N}{an integer, number of coordinates, see package description}

\item{c}{a positive number, penalty for model selection rule, see package description}

\item{B}{an integer specifying the number of replicates used in p-value computation}

\item{compute.p}{a logical value indicating whether to compute a p-value}
}
\description{
Performs data driven smooth test for the classical two- and k-sample problems as described in details in Wylupek (2010).
}
\examples{
# H0 is true
x = runif(80)
y = runif(80)
z = runif(80)
ddst.ksample.test(x, y, z, compute.p=TRUE)
ddst.ksample.test(list(x, y, z), compute.p=TRUE)

# known fixed alternative
x = runif(80)
y = rbeta(80,4,2)
w = rnorm(30)
z = rexp(10, 1)
ddst.ksample.test(x, y, w, z, compute.p=TRUE)
ddst.ksample.test(list(x, y, w, z), compute.p=TRUE)
}
\keyword{htest}
