% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddst.twosample.test.R
\name{ddst.twosample.test}
\alias{ddst.twosample.test}
\title{Data Driven Smooth Test for Two-Sample Problem}
\usage{
ddst.twosample.test(
  x,
  y,
  d.N = 12,
  c = 2,
  B = 1e+05,
  compute.p = TRUE,
  alpha = 0.05,
  compute.cv = TRUE
)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data}

\item{y}{a (non-empty) numeric vector of data}

\item{d.N}{an integer specifying the maximum dimension considered, only for advanced users}

\item{c}{a calibrating parameter in the penalty in the model selection rule}

\item{compute.p}{a logical value indicating whether to compute a p-value or not}

\item{alpha}{a significance level}

\item{compute.cv}{a logical value indicating whether to compute a critical value corresponding to the significance level alpha or not}

\item{nr}{an integer specifying the number of runs for a p-value and a critical value computation if any}
}
\description{
Performs data driven smooth test for the classical two-sample problem.
It is a special case of data driven test for k-samples.
Detailed description of the test statistic is provided in Wylupek (2010).
}
\examples{
set.seed(7)
# H0 is true
x <- runif(80)
y <- runif(80)
t <- ddst.twosample.test(x, y)
t
plot(t)

# H0 is false
x <- runif(80)
y <- rexp(80, 1)
t <- ddst.twosample.test(x, y)
t
plot(t)

}
\references{
Data-driven k-sample tests. Wylupek (2010) \url{https://www.jstor.org/stable/40586684?seq=1}
}
\keyword{htest}
