% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddst.umbrella.test.R
\name{ddst.umbrella.test}
\alias{ddst.umbrella.test}
\title{Data Driven Test for k-Sample Umbrella Alternatives}
\usage{
ddst.umbrella.test(x, ..., tlh.p = 2.2, tl.n = 2.2, p = 3,
  type = "T", B = 100, compute.p = FALSE)
}
\arguments{
\item{x}{a list with vectors or vectors (in thos case other vectors will be in ...)}

\item{...}{other vectors}

\item{B}{an integer specifying the number of replicates used in p-value computation}

\item{compute.p}{a logical value indicating whether to compute a p-value}

\item{d_N}{an integer, number of coordinates, see package description}

\item{c}{a positive number, penalty for model selection rule, see package description}
}
\description{
Performs data driven smooth test for so-called umbrella alternatives in k-sample problem;
see Wylupek (2016).
}
\examples{
# H0 is true
x = runif(80)
y = runif(80)
z = runif(80)
ddst.umbrella.test(x, y, z, compute.p=TRUE)
ddst.umbrella.test(x, y, z, compute.p=TRUE, type = "S")
ddst.umbrella.test(x, y, z, compute.p=TRUE, type = "M")
ddst.umbrella.test(list(x, y, z), compute.p=TRUE)

# known fixed alternative
x = runif(80)
y = rbeta(80,4,2)
w = rnorm(30)
z = rexp(10, 1)
ddst.umbrella.test(x, y, w, z)
ddst.umbrella.test(x, y, w, z, type = "S")
ddst.umbrella.test(x, y, w, z, type = "M")
ddst.umbrella.test(list(x, y, w, z))
}
\keyword{htest}
