% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddst.exp.test.R
\name{ddst.exp.test}
\alias{ddst.exp.test}
\alias{ddst.exp.Nk}
\title{Data Driven Smooth Test for Exponentiality}
\usage{
ddst.exp.test(
  x,
  base = ddst.base.legendre,
  d.n = 10,
  c = 100,
  nr = 1e+05,
  compute.p = TRUE,
  alpha = 0.05,
  compute.cv = TRUE,
  ...
)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values}

\item{base}{a function which returns an orthonormal system, possible choice: \code{ddst.base.legendre} for the Legendre polynomials and \code{ddst.base.cos} for the cosine system}

\item{d.n}{an integer specifying the maximum dimension considered, only for advanced users}

\item{c}{a calibrating parameter in the penalty in the model selection rule}

\item{nr}{an integer specifying the number of runs for a p-value and a critical value computation if any}

\item{compute.p}{a logical value indicating whether to compute a p-value or not}

\item{alpha}{a significance level}

\item{compute.cv}{a logical value indicating whether to compute a critical value corresponding to the significance level alpha or not}

\item{...}{further arguments}
}
\value{
An object of class \code{htest}
\item{statistic }{the value of the test statistic.}
\item{parameter }{the number of choosen coordinates (k).}
\item{method }{a character string indicating the parameters of performed test. }
\item{data.name }{a character string giving the name(s) of the data. }
\item{p.value }{the p-value for the test, computed only if \code{compute.p=T}.}
}
\description{
Performs data driven smooth test for composite hypothesis of exponentiality.
Null density is given by \eqn{f(z;gamma) = exp(-z/gamma)}  for z >= 0 and 0 otherwise.
Modelling alternatives similarly as in Kallenberg and Ledwina (1997 a,b).
}
\examples{
set.seed(7)
# H0 is true
z <- rexp(80,4)
\dontrun{
t <- ddst.exp.test (z, compute.p = TRUE, d.n = 10)
t
plot(t)

# H0 is false
z = rchisq(80,4)
(t = ddst.exp.test (z, compute.p = TRUE, d.n = 10))
t$p.value
plot(t)
}
}
\references{
Kallenberg, W.C.M., Ledwina, T. (1997 a). Data driven smooth tests for composite hypotheses: Comparison of powers. \eqn{ J. Statist. Comput. Simul.} \bold{ 59}, 101--121.

Kallenberg, W.C.M.,  Ledwina, T. (1997 b). Data driven smooth tests when the hypothesis is composite. \eqn{ J. Amer. Statist. Assoc.} \bold{ 92}, 1094--1104.
}
\keyword{htest}
