% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddst.stochasticorder.test.R
\name{ddst.forstochdom.test}
\alias{ddst.forstochdom.test}
\title{Data Driven Nonparametric Test for Stochastic Ordering}
\usage{
ddst.forstochdom.test(x, y, t = 2.2, d = 4)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values}

\item{y}{a (non-empty) numeric vector of data values}

\item{t}{a positive number, penalty for model selection rule, see package description}

\item{d}{an integer, number of coordinates that measure potential deviation from null hypothesis}

\item{...}{further arguments}
}
\description{
Performs the data driven smooth test for detection of
the stochastic ordering, as described
in detail in Ledwina and Wyłupek (2012).
Suppose that we have random samples from two distributions F and G.
The null hypothesis is that F(x) >= G(x) for all x while the alternative is that at
F(x) < G(x) for some x.
Detailed description of the test statistic is provided in Ledwina and Wylupek (2012).
}
\examples{
set.seed(13)
library("rmutil", warn.conflicts = FALSE)
# 1. Pareto(1)/Pareto(1.5)
#uzyc parametrow z tabeli 3, p. 742, np dla m = n = 50,
x <- rpareto(50, 2, 2)
y <- rpareto(50, 1.5, 1.5)
t <- ddst.forstochdom.test(x, y)
t
plot(t)

# 2. Laplace(0,1)/Laplace(1,25)
x <- rlaplace(50, 0, 1)
y <- rlaplace(50, 1, 25)
t <- ddst.forstochdom.test(x, y)
t
plot(t)

# 3. LN(0.85,0.6)/LN(1.2,0.2)
x <- rlnorm(50, 0.85, 0.6)
y <- rlnorm(50, 1.2, 0.2)
t <- ddst.forstochdom.test(x, y)
t
plot(t)

\dontrun{
# Generate distribution of test statistic
N <- 1000
samp <- replicate(N, {
   x <- runif(30)
   y <- runif(30)
   # statistics with Schwartz penalty
   ddst.forstochdom.test(x, y)$statistic
})
quantile(samp, 0.95)
plot(ecdf(samp))
}

}
\references{
Nonparametric tests for stochastic ordering. Ledwina and Wyłupek (2012) \url{https://doi.org/10.1007/s11749-011-0278-7}
}
\keyword{htest}
