% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddst.ksample.test.R
\name{ddst.ksample.test}
\alias{ddst.ksample.test}
\title{Data Driven k-Sample Test}
\usage{
ddst.ksample.test(x, ..., d_N = 12, c = 2.3)
}
\arguments{
\item{x}{a list with k-vectors or a single vector.}

\item{...}{if x is a single vector, then remaing k-1 vectors are provided in the ... argument}

\item{d_N}{an integer, number of coordinates that measure potential deviation from null hypothesis}

\item{c}{a positive number, penalty for model selection rule. Section 5.1 in Wylupek (2010), suggest that a good choice is c = 2, when k = 2, and  c = 2.3, when k >= 3.}
}
\description{
Performs data driven smooth test for the classical k-sample problem.
Suppose that we have random samples from k distributions F_i where i = 1, ..., k.
The null hypothesis is that F_1 = ... = F_k while the alternative is that at
least two distributions are different.
Detailed description of the test statistic is provided in Wylupek (2010).
}
\examples{
# H0 is true
x = runif(80)
y = runif(80)
z = runif(80)
t <- ddst.ksample.test(x, y, z)
t <- ddst.ksample.test(list(x, y, z))
t
plot(t)

# H0 is false
x = runif(80)
y = rexp(80, 1)
z = runif(80)
t <- ddst.ksample.test(x, y, z)
t
plot(t)

}
\references{
Data-driven k-sample tests. Wylupek (2010) \url{https://www.jstor.org/stable/40586684?seq=1}
}
\keyword{htest}
