% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddst.normbounded.test.R
\name{ddst.normbounded.test}
\alias{ddst.normbounded.test}
\alias{ddst.norm.Nk}
\title{Data Driven Smooth Test for Normality - Bounded Basis Functions}
\usage{
ddst.normbounded.test(x, base = ddst.base.legendre, c = 100, B = 100,
  compute.p = FALSE, Dmax = 5, ...)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values}

\item{base}{a function which returns orthogonal system, might be \code{ddst.base.legendre} for Legendre polynomials
or \code{ddst.base.cos} for cosine system, see package description}

\item{c}{a parameter for model selection rule, see package description}

\item{B}{an integer specifying the number of replicates used in p-value computation}

\item{compute.p}{a logical value indicating whether to compute a p-value}

\item{Dmax}{an integer specifying the maximum number of coordinates, only for advanced users}

\item{...}{further arguments}
}
\value{
An object of class \code{htest}
\item{statistic }{the value of the test statistic.}
\item{parameter }{the number of choosen coordinates (k).}
\item{method }{a character string indicating the parameters of performed test. }
\item{data.name }{a character string giving the name(s) of the data. }
\item{p.value }{the p-value for the test, computed only if \code{compute.p = TRUE}.}
}
\description{
Performs data driven smooth test for composite hypothesis of normality
Null density is given by
\eqn{
  f(z;\gamma)=1/(\sqrt{2 \pi}\gamma_2) \exp(-(z-\gamma_1)^2/(2 \gamma_2^2))} for \eqn{z \in R}.
We model alternatives similarly as in Kallenberg and Ledwina (1997 a,b) using Legendre's polynomials or cosine basis.
For more details see: \url{http://www.biecek.pl/R/ddst/description.pdf}.
}
\examples{
# for given vector of 19 numbers
z <- c(13.41, 6.04, 1.26, 3.67, -4.54, 2.92, 0.44, 12.93, 6.77, 10.09,
    4.10, 4.04, -1.97, 2.17, -5.38, -7.30, 4.75, 5.63, 8.84)
ddst.norm.test(z, compute.p=TRUE)

# H0 is true
z <- rnorm(80)
t <- ddst.norm.test(z, compute.p=TRUE)
t
plot(t)

# H0 is false
z <- rexp(80,4)
t <- ddst.norm.test(z, B=5000, compute.p=TRUE)
t
plot(t)

# for Tephra data
z <- c(-1.748789, -1.75753, -1.740102, -1.740102, -1.731467, -1.765523,
      -1.761521, -1.72522, -1.80371, -1.745624, -1.872957, -1.729121,
      -1.81529, -1.888637, -1.887761, -1.881645, -1.91518, -1.849769,
      -1.755141, -1.665687, -1.764721, -1.736171, -1.736956, -1.737742,
      -1.687537, -1.804534, -1.790593, -1.808661, -1.784081, -1.729903,
      -1.711263, -1.748789, -1.772755, -1.72756, -1.71358, -1.821116,
      -1.839588, -1.839588, -1.830321, -1.807835, -1.747206, -1.788147,
      -1.759923, -1.786519, -1.726779, -1.738528, -1.754345, -1.781646,
      -1.641949, -1.755936, -1.775175, -1.736956, -1.705103, -1.743255,
      -1.82613, -1.826967, -1.780025, -1.684504, -1.751168)
t <- ddst.norm.test(z, compute.p=TRUE)
t
plot(t)
}
\references{
Chen, L., Shapiro, S.S. (1995). An alternative test for normality based on normalized spacings. \emph{ J. Statist. Comput. Simulation} 53, 269--288.

Inglot, T., Ledwina, T. (2006). Towards data driven selection of a penalty function for data driven Neyman tests. \emph{ Linear Algebra and its Appl.} \bold{ 417}, 579--590.

Janic, A. and Ledwina, T. (2008). Data-driven tests for a location-scale family revisited. \emph{ J. Statist. Theory. Pract. Special issue on Modern Goodness of Fit Methods.}.

Kallenberg, W.C.M., Ledwina, T. (1997 a). Data driven smooth tests for composite hypotheses: Comparison of powers. \emph{ J. Statist. Comput. Simul.} \bold{ 59}, 101--121.

Kallenberg, W.C.M.,  Ledwina, T. (1997 b). Data driven smooth tests when the hypothesis is composite. \emph{ J. Amer. Statist. Assoc.} \bold{ 92}, 1094--1104.
}
\keyword{htest}
