% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddst.stochasticdominance.test.R
\name{ddst.stochasticdominance.test}
\alias{ddst.stochasticdominance.test}
\title{Data Driven Two-Sample Test Against One-Sided Alternatives}
\usage{
ddst.stochasticdominance.test(x, y, t = 2.2, k.N = 4, B = 1000,
  compute.p = FALSE, alpha = 0.05, ...)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values}

\item{y}{a (non-empty) numeric vector of data values}

\item{t}{a positive number, penalty for model selection rule, see package description}

\item{k.N}{(TODO: is it D ?)}

\item{B}{an integer specifying the number of replicates used in p-value computation}

\item{compute.p}{a logical value indicating whether to compute a p-value}

\item{...}{further arguments}
}
\description{
Performs data driven smooth non-parametric two-sample test against one-sided alternatives
(stochastic dominance)
}
\examples{
# H0 is true
x = runif(80)
y = runif(80)
ddst.twosample.test(x, y, compute.p=TRUE)

# known fixed alternative
x = runif(80)
y = rbeta(80,4,2)
ddst.twosample.test(x, y, compute.p=TRUE)
}
\keyword{htest}
