% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddst.stochasticorder.test.R
\name{ddst.stochasticorder.test}
\alias{ddst.stochasticorder.test}
\title{Data Driven Nonparametric Test for Stochastic Ordering}
\usage{
ddst.stochasticorder.test(x, y, t = 2.2, k.N = 4)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values}

\item{y}{a (non-empty) numeric vector of data values}

\item{t}{tuning parameter, see test description}

\item{k.N}{The choice of k(N) belongs to the user and should reflect the individual aims and needs. Some authors’ recommendation is given in Section 5.2.}
}
\description{
Performs data driven smooth non-parametric two-sample test for stochastic ordering
}
\details{
This test is described in
Nonparametric tests for stochastic ordering
Ledwina, T. and Wyłupek, G. TEST (2012) 21: 730.
https://doi.org/10.1007/s11749-011-0278-7
}
\examples{
# H0 is true
x = runif(80)
y = runif(80)
ddst.stochasticorder.test(x, y)

}
\keyword{htest}
