% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddst.upward.test.R
\name{ddst.upwardtrend.test}
\alias{ddst.upwardtrend.test}
\title{Data Driven Test for Upward Trend}
\usage{
ddst.upwardtrend.test(x, ..., tlh.p = 2.2, tl.n = 2.2)
}
\arguments{
\item{x}{a list with k-vectors or a single vector.}

\item{...}{if x is a single vector, then remaing k-1 vectors are provided in the ... argument}

\item{tlh.p}{a positive number, penalty for model selection rule}

\item{tl.n}{a positive number, penalty for model selection rule}
}
\description{
Performs data driven smooth test for upward trend in k-sample problem.
Suppose that we have random samples from k distributions F_i where i = 1, ..., k.
The null hypothesis is that there is lack of trend,
i.e. F_1 >= ... >= F_k and F_i != F_j for some i and j.
The alternative is that there is a trend
i.e. F_1 >= ... >= F_k and F_i != F_j for some i and j.
This test is implemented as a special case of an umbrella test.
}
\examples{
# H0 is true
x = runif(80)
y = runif(80) + 0.2
z = runif(80) + 0.4
t <- ddst.upwardtrend.test(list(x, y, z))
t
plot(t)

# known fixed alternative
x1 = rnorm(80)
x2 = rnorm(80) + 2
x3 = rnorm(80) + 4
x4 = rnorm(80) + 3
t <- ddst.upwardtrend.test(list(x1, x2, x3, x4))
t
plot(t)

}
\references{
An automatic test for the umbrella alternatives. Wylupek (2016) \url{https://onlinelibrary.wiley.com/doi/abs/10.1111/sjos.12231}
}
\keyword{htest}
