% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddst.plot.R
\name{plot.ddst.test}
\alias{plot.ddst.test}
\title{Plot Function fo Data Driven Tests}
\usage{
\method{plot}{ddst.test}(x, ...)
}
\arguments{
\item{x}{result from ddst test function}

\item{...}{currently not used}
}
\description{
Plots coordinates for selected test statistics
}
\examples{
# H0 is true
x <- runif(80)
y <- runif(80)
z <- runif(80)
t <- ddst.ksample.test(list(x, y, z))
plot(t)
}
